closed #13 total 값 long으로 변

+++ src/main/java/org/l3cache/dao/PostManager.java
@@ -103,7 +103,7 @@ public class PostManager {
 		return sqlSession.selectList("PostMapper.selectUserPostsList", postSel);
 	}
 
-	public int getUserPostsCount(int uid) {
+	public long getUserPostsCount(int uid) {
 		return sqlSession.selectOne("PostMapper.countUserPostsList", uid);
 	}
